Add `niceargs` internal package
I was fighting with Go's built-in `flag` package. With `flag` you can
build very nice command line programs that take arguments like this:

    command subcommand --flag flag-value argument

But, as an end user, you can't enter arguments like this:

    command subcommand argument --flag flag-value

This wasn't a deal-breaker for me, but it seems very normal to let users
enter commands this way. My naive solution is to just reorder the
user-given arguments using `niceargs.List()`.